16) Bill Managing Agent/requirements.txt
streamlit
python-dotenv
google-generativeai
pyautogen